requirements.txt
boto3==1.36.11
PyYAML==6.0.2